# backend/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/comments": {
      "post": {
        "operationId": "CommentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Commentaire créé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          },
          "404": {
            "description": "Échéance non trouvée"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Créer un nouveau commentaire",
        "tags": [
          "comments"
        ]
      },
      "get": {
        "operationId": "CommentsController_findAll",
        "parameters": [
          {
            "name": "authorId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deadlineId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "visibility",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "privé",
                "public"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des commentaires",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Comment"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer tous les commentaires",
        "tags": [
          "comments"
        ]
      }
    },
    "/comments/{id}": {
      "get": {
        "operationId": "CommentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID du commentaire",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Commentaire trouvé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          },
          "404": {
            "description": "Commentaire non trouvé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer un commentaire par son ID",
        "tags": [
          "comments"
        ]
      },
      "patch": {
        "operationId": "CommentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID du commentaire",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Commentaire mis à jour",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          },
          "403": {
            "description": "Accès non autorisé"
          },
          "404": {
            "description": "Commentaire non trouvé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mettre à jour un commentaire",
        "tags": [
          "comments"
        ]
      },
      "delete": {
        "operationId": "CommentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID du commentaire",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Commentaire supprimé"
          },
          "403": {
            "description": "Accès non autorisé"
          },
          "404": {
            "description": "Commentaire non trouvé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprimer un commentaire",
        "tags": [
          "comments"
        ]
      }
    },
    "/comments/deadline/{deadlineId}": {
      "get": {
        "operationId": "CommentsController_findByDeadline",
        "parameters": [
          {
            "name": "deadlineId",
            "required": true,
            "in": "path",
            "description": "ID de l'échéance",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des commentaires de l'échéance",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Comment"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer les commentaires par échéance",
        "tags": [
          "comments"
        ]
      }
    },
    "/comments/author/{authorId}": {
      "get": {
        "operationId": "CommentsController_findByAuthor",
        "parameters": [
          {
            "name": "authorId",
            "required": true,
            "in": "path",
            "description": "ID de l'auteur",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des commentaires de l'auteur",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Comment"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer les commentaires par auteur",
        "tags": [
          "comments"
        ]
      }
    },
...
  },
  "info": {
    "title": "API de Gestion d'Échéances",
    "description": "API pour l'application de gestion d'échéances du secteur défense",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "Contenu du commentaire",
            "example": "Nous devons avancer sur cette tâche rapidement."
          },
          "authorId": {
            "type": "string",
            "description": "ID de l'utilisateur auteur du commentaire",
            "example": "5f8d0f55-e477-4915-b522-c21ba1f70d9e"
          },
          "deadlineId": {
            "type": "string",
            "description": "ID de l'échéance associée au commentaire",
            "example": "7a9d0f35-d562-4219-c732-d35cf1a2e5f9"
          },
          "visibility": {
            "type": "string",
            "description": "Visibilité du commentaire",
            "enum": [
              "privé",
              "public"
            ],
            "example": "public"
          }
        },
        "required": [
          "content",
          "authorId",
          "deadlineId",
          "visibility"
        ]
      },
      "Comment": {
        "type": "object",
        "properties": {}
      },
      "UpdateCommentDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "Contenu du commentaire",
            "example": "Nous devons avancer sur cette tâche rapidement."
          },
          "authorId": {
            "type": "string",
            "description": "ID de l'utilisateur auteur du commentaire",
            "example": "5f8d0f55-e477-4915-b522-c21ba1f70d9e"
          },
          "deadlineId": {
            "type": "string",
            "description": "ID de l'échéance associée au commentaire",
            "example": "7a9d0f35-d562-4219-c732-d35cf1a2e5f9"
          },
          "visibility": {
            "type": "string",
            "description": "Visibilité du commentaire",
            "enum": [
              "privé",
              "public"
            ],
            "example": "public"
          }
        }
      },
...
    }
  }
}